+++ src/app/review/write/_components/ReviewImg.tsx
@@ -8,6 +8,9 @@ type ReviewImgProps = {
 };
 
 export default function ReviewImg({ urlList, setList }: ReviewImgProps) {
+  const MAX_IMAGE_COUNT = 2;
+  const additionalImages = urlList.slice(0, MAX_IMAGE_COUNT);
+  const extraCount = urlList.length - MAX_IMAGE_COUNT - 1;
   const reviewStatus = reviewStore();
 
   const uploadImgHandler = (img: string) => {
@@ -18,7 +21,7 @@ export default function ReviewImg({ urlList, setList }: ReviewImgProps) {
 
   return (
     <ul className="flex justify-center h-28 max:h-[90px] overflow-hidden gap-3 my-3">
-      {urlList.map((img, idx) => (
+      {additionalImages.map((img, idx) => (
         <li key={idx} className="relative basis-1/3 rounded-md overflow-hidden">
           <Image
             src={img}
@@ -36,6 +39,30 @@ export default function ReviewImg({ urlList, setList }: ReviewImgProps) {
           </span>
         </li>
       ))}
+      {urlList.length > 2 && (
+        <li className="relative basis-1/3 rounded-md overflow-hidden">
+          <Image
+            src={urlList[2]}
+            alt={urlList[2]}
+            fill={true}
+            // Todo: 이미지 사이즈 최적화하기
+            sizes="(max-width: 732px) 90vw, (max-width: 992px) 45vw, 320px"
+            className="object-cover"
+            priority
+          />
+          {extraCount > 0 ? (
+            <span className="absolute z-modal left-0 top-0 w-full h-full bg-black opacity-50 text-white flex justify-center items-center">
+              +{extraCount}
+            </span>
+          ) : (
+            <span
+              onClick={() => uploadImgHandler(urlList[2])}
+              className="cursor-pointer absolute top-3 right-1 z-float w-5 h-5 bg-button_bg_default hover:bg-button_bg_clicked text-white rounded-full flex items-center justify-center">
+              x
+            </span>
+          )}
+        </li>
+      )}
     </ul>
   );
 }
